+++ editor/client/src/layouts/full-page/FullPage.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="layout--full-page">
+    <router-view></router-view>
+  </div>
+</template>