+++ editor/client/src/router.js
@@ -1,0 +1,84 @@
+import Vue from 'vue'
+import Router from 'vue-router'
+
+Vue.use(Router)
+
+const router = new Router({
+  mode: 'history',
+  base: process.env.BASE_URL,
+  scrollBehavior () {
+    return { x: 0, y: 0 }
+  },
+  routes: [
+
+    {
+    // =============================================================================
+    // MAIN LAYOUT ROUTES
+    // =============================================================================
+      path: '',
+      component: () => import('./layouts/main/Main.vue'),
+      children: [
+        // =============================================================================
+        // Theme Routes
+        // =============================================================================
+        {
+          path: '/',
+          //redirect: '/dashboard/restqa'
+          redirect: '/pages/comingsoon'
+        },
+        {
+          path: '/dashboard/restqa',
+          name: 'dashboard-restgqa',
+          component: () => import('./views/DashboardRestQA.vue'),
+          meta: {
+            rule: 'editor'
+          }
+        }
+      ]
+    },
+    // =============================================================================
+    // FULL PAGE LAYOUTS
+    // =============================================================================
+    {
+      path: '',
+      component: () => import('@/layouts/full-page/FullPage.vue'),
+      children: [
+        {
+          path: '/pages/comingsoon',
+          name: 'page-coming-soon',
+          component: () => import('@/views/pages/ComingSoon.vue'),
+          meta: {
+            rule: 'editor'
+          }
+        },
+        {
+          path: '/pages/error-404',
+          name: 'page-error-404',
+          component: () => import('@/views/pages/Error404.vue'),
+          meta: {
+            rule: 'editor'
+          }
+        }
+      ]
+    },
+    // Redirect to 404 page, if no match found
+    {
+      path: '*',
+      redirect: '/pages/error-404'
+    }
+  ]
+})
+
+router.afterEach(() => {
+  // Remove initial loading
+  const appLoading = document.getElementById('loading-bg')
+  if (appLoading) {
+    appLoading.style.display = 'none'
+  }
+})
+
+router.beforeEach((to, from, next) => {
+  next()
+})
+
+export default router