+++ editor/client/src/views/DashboardRestQA.vue
@@ -1,0 +1,19 @@
+<template>
+  <div id="dashboard-analytics"></div>
+</template>
+
+<script>
+
+export default {
+  data () {
+    return {}
+  },
+  components: {
+  },
+  created () {
+  }
+}
+</script>
+
+<style lang="scss">
+</style>